move error boundary to root
