+++ app/Components/Card.jsx
@@ -17,20 +17,6 @@ export default function Card({ title, value, color, onClick }) {
   )
 }
 
-export function ErrorBoundary({ error }) {
-  return (
-    <div className="ErrorBoundary">
-      <h1 className="ErrorBoundary__Text">Something went wrong</h1>
-      <p className="ErrorBoundary__Text">{error.message}</p>
-    </div>
-  )
-}
-
-// props validation
-ErrorBoundary.propTypes = {
-  error: PropTypes.object.isRequired,
-}
-
 Card.propTypes = {
   title: PropTypes.string.isRequired,
   value: PropTypes.string.isRequired,
