+++ app/Components/SearchBar.jsx
@@ -38,18 +38,4 @@ SearchBar.propTypes = {
   clearSearch: PropTypes.func.isRequired,
 }
 
-export function ErrorBoundary({ error }) {
-  return (
-    <div className="ErrorBoundary">
-      <h1 className="ErrorBoundary__Text">Something went wrong</h1>
-      <p className="ErrorBoundary__Text">{error.message}</p>
-    </div>
-  )
-}
-
-// props validation
-ErrorBoundary.propTypes = {
-  error: PropTypes.object.isRequired,
-}
-
 export default SearchBar
