+++ app/Components/Transaction.jsx
@@ -42,17 +42,3 @@ Transaction.propTypes = {
   date: PropTypes.string.isRequired,
   note: PropTypes.string.isRequired,
 }
-
-export function ErrorBoundary({ error }) {
-  return (
-    <div className="ErrorBoundary">
-      <h1 className="ErrorBoundary__Text">Something went wrong</h1>
-      <p className="ErrorBoundary__Text">{error.message}</p>
-    </div>
-  )
-}
-
-// props validation
-ErrorBoundary.propTypes = {
-  error: PropTypes.object.isRequired,
-}
