+++ app/Components/dropdown.jsx
@@ -18,20 +18,6 @@ export default function Dropdown({ title, options, setCategory, defaultValue, is
   )
 }
 
-export function ErrorBoundary({ error }) {
-  return (
-    <div className="ErrorBoundary">
-      <h1 className="ErrorBoundary__Text">Something went wrong</h1>
-      <p className="ErrorBoundary__Text">{error.message}</p>
-    </div>
-  )
-}
-
-// props validation
-ErrorBoundary.propTypes = {
-  error: PropTypes.object.isRequired,
-}
-
 // default props
 Dropdown.defaultProps = {
   title: "",
